% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.density.R
\name{ST.density}
\alias{ST.density}
\title{Illustrate the spatial distribution of gene expression}
\usage{
ST.density(ST.object, Gene, Threshold, Path)
}
\arguments{
\item{ST.object}{Seurat object which was pre-processed and the dimentionality reduction has been done}

\item{Gene}{The desired gene to investigate the spatial distribution}

\item{Threshold}{The limit ratio that you would like to apply to see the spots with expression higher than this threshold}

\item{Path}{Path to the Spatial folder in output directory of SpaceRanger which contains the coordination in CSV format}
}
\value{
A density ggplot illustration of the spatial distribution of gene expression
}
\description{
The package requires four parameters including a) processed Seurat object, b) desired gene
c) Ratio threshold for the expression of gene in each spot d) Path to the Spatial folder in
SpaceRanger output directory. As a results, it illustrates the spatial distribution of desired gene
in tissue section as a density plot.
}
